refactor: MVC 패턴으로 리팩터링
