+++ app/(home)/_components/Filtering.tsx
@@ -2,31 +2,18 @@
 
 import { Button } from "@/components/ui/button";
 import { ChevronDownIcon } from "lucide-react";
-import { useState } from "react";
-import { FilterType, FilterValue } from "../type";
+import { useFilteringController } from "../controllers/FilteringController";
 import FilteringModal from "./FilteringModal";
 
 export default function Filtering() {
-  // 모달 열림/닫힘 상태 관리
-  const [open, setOpen] = useState(false);
-  // 현재 선택된 필터 타입 관리
-  const [filterType, setFilterType] = useState<FilterType>("정렬");
-  // 각 섹션별 선택된 필터 상태 관리
-  const [selectedFilters, setSelectedFilters] = useState<
-    Record<FilterType, FilterValue>
-  >({
-    정렬: "",
-    촬영시기: "",
-    스타일: [],
-    패키지: "",
-    가격: 0,
-  });
-
-  // 필터 버튼 클릭 핸들러
-  const handleFilterClick = (type: FilterType) => {
-    setFilterType(type);
-    setOpen(true);
-  };
+  const {
+    open,
+    filterType,
+    selectedFilters,
+    handleFilterClick,
+    setOpen,
+    setSelectedFilters,
+  } = useFilteringController();
 
   return (
     <>
